feat: Add CLI and GUI modules for toggling taskbar visibility
This commit adds two new modules, `cli.py` and `gui.py`, to the project. The `cli.py` module contains functions for parsing command-line arguments and changing the hotkey used to toggle the taskbar visibility. The `gui.py` module is currently empty but will be used for implementing a graphical user interface in future updates.

The main module (`main.py`) has been updated to import the new modules and call their respective main functions based on a condition. Additionally, the function for toggling the taskbar visibility has been moved to a separate `utils.py` module.

These changes provide more flexibility by allowing users to toggle the taskbar visibility either through command-line arguments or a graphical user interface.
