+++ README.md
@@ -25,4 +25,3 @@ Here's how other solutions stack up:
 ## Usage
 
 (Include usage instructions, examples, and/or code snippets here)
-e
